# feat(optimizer): Add Lance count() pushdown optimization