--- a/src/common/scan-info/src/pushdowns.rs
+++ b/src/common/scan-info/src/pushdowns.rs
@@ -28,11 +28,16 @@ pub struct Pushdowns {
     /// The `filters` field is kept for backward compatibility;
     /// it represents all current filters.
     pub pushed_filters: Option<Vec<ExprRef>>,
+
+    // /// Optional aggregation pushdown.
+    /// This is used to indicate that the scan operator can perform an aggregation.
+    /// This is useful for scans that can perform aggregations like `count`
+    pub aggregation: Option<ExprRef>,
 }
 
 impl Default for Pushdowns {
     fn default() -> Self {
-        Self::new(None, None, None, None, None)
+        Self::new(None, None, None, None, None, None)
     }
 }
 
@@ -44,6 +49,7 @@ impl Pushdowns {
         columns: Option<Arc<Vec<String>>>,
         limit: Option<usize>,
         sharder: Option<Sharder>,
+        aggregation: Option<ExprRef>,
     ) -> Self {
         Self {
             filters,
@@ -52,6 +58,7 @@ impl Pushdowns {
             limit,
             sharder,
             pushed_filters: None,
+            aggregation,
         }
     }
 
@@ -72,6 +79,7 @@ impl Pushdowns {
             limit,
             sharder: self.sharder.clone(),
             pushed_filters: self.pushed_filters.clone(),
+            aggregation: self.aggregation.clone(),
         }
     }
 
@@ -84,6 +92,7 @@ impl Pushdowns {
             limit: self.limit,
             sharder: self.sharder.clone(),
             pushed_filters: self.pushed_filters.clone(),
+            aggregation: self.aggregation.clone(),
         }
     }
 
@@ -96,6 +105,7 @@ impl Pushdowns {
             limit: self.limit,
             sharder: self.sharder.clone(),
             pushed_filters: self.pushed_filters.clone(),
+            aggregation: self.aggregation.clone(),
         }
     }
 
@@ -108,6 +118,7 @@ impl Pushdowns {
             limit: self.limit,
             sharder: self.sharder.clone(),
             pushed_filters: self.pushed_filters.clone(),
+            aggregation: self.aggregation.clone(),
         }
     }
 
@@ -120,6 +131,7 @@ impl Pushdowns {
             limit: self.limit,
             sharder,
             pushed_filters: self.pushed_filters.clone(),
+            aggregation: self.aggregation.clone(),
         }
     }
 
@@ -132,6 +144,20 @@ impl Pushdowns {
             limit: self.limit,
             sharder: self.sharder.clone(),
             pushed_filters,
+            aggregation: self.aggregation.clone(),
+        }
+    }
+
+    #[must_use]
+    pub fn with_aggregation(&self, aggregation: Option<ExprRef>) -> Self {
+        Self {
+            filters: self.filters.clone(),
+            partition_filters: self.partition_filters.clone(),
+            columns: self.columns.clone(),
+            limit: self.limit,
+            sharder: self.sharder.clone(),
+            pushed_filters: self.pushed_filters.clone(),
+            aggregation,
         }
     }
 
@@ -152,6 +178,9 @@ impl Pushdowns {
         }
         if let Some(sharder) = &self.sharder {
             res.push(format!("Sharder = {sharder}"));
+        }
+        if let Some(aggregation) = &self.aggregation {
+            res.push(format!("Aggregation pushdown = {aggregation}"));
         }
         res
     }
@@ -186,6 +215,9 @@ impl DisplayAs for Pushdowns {
                 }
                 if let Some(sharder) = &self.sharder {
                     sub_items.push(format!("sharder: {sharder}"));
+                }
+                if let Some(aggregation) = &self.aggregation {
+                    sub_items.push(format!("aggregation: {aggregation}"));
                 }
                 s.push_str(&sub_items.join(", "));
                 s.push('}');