--- a/src/common/scan-info/src/python.rs
+++ b/src/common/scan-info/src/python.rs
@@ -3,7 +3,8 @@ use pyo3::prelude::*;
 pub mod pylib {
     use std::sync::Arc;
 
-    use daft_dsl::python::PyExpr;
+    use daft_core::count_mode::CountMode;
+    use daft_dsl::{python::PyExpr, AggExpr, Expr};
     use daft_schema::python::field::PyField;
     use pyo3::{exceptions::PyAttributeError, prelude::*, pyclass};
     use serde::{Deserialize, Serialize};
@@ -164,12 +165,14 @@ pub mod pylib {
             partition_filters = None,
             columns = None,
             limit = None,
+            aggregation = None,
         ))]
         pub fn new(
             filters: Option<PyExpr>,
             partition_filters: Option<PyExpr>,
             columns: Option<Vec<String>>,
             limit: Option<usize>,
+            aggregation: Option<PyExpr>,
         ) -> Self {
             let pushdowns = Pushdowns::new(
                 filters.map(|f| f.expr),
@@ -177,6 +180,7 @@ pub mod pylib {
                 columns.map(Arc::new),
                 limit,
                 None,
+                aggregation.map(|f| f.expr),
             );
             Self(Arc::new(pushdowns))
         }
@@ -212,11 +216,37 @@ pub mod pylib {
             self.0.columns.as_deref().cloned()
         }
 
+        #[getter]
+        #[must_use]
+        pub fn aggregation(&self) -> Option<PyExpr> {
+            self.0
+                .aggregation
+                .as_ref()
+                .map(|e| PyExpr { expr: e.clone() })
+        }
+
         pub fn filter_required_column_names(&self) -> Option<Vec<String>> {
             self.0
                 .filters
                 .as_ref()
                 .map(daft_dsl::optimization::get_required_columns)
+        }
+
+        pub fn aggregation_required_column_names(&self) -> Option<Vec<String>> {
+            self.0
+                .aggregation
+                .as_ref()
+                .map(daft_dsl::optimization::get_required_columns)
+        }
+
+        pub fn aggregation_count_mode(&self) -> Option<CountMode> {
+            match self.0.aggregation.as_ref() {
+                Some(expr) => match expr.as_ref() {
+                    Expr::Agg(AggExpr::Count(_, count_mode)) => Some(*count_mode),
+                    _ => None,
+                },
+                None => None,
+            }
         }
     }
 }