--- a/src/daft-logical-plan/src/optimization/rules/push_down_projection.rs
+++ b/src/daft-logical-plan/src/optimization/rules/push_down_projection.rs
@@ -1167,6 +1167,7 @@ mod tests {
                 filters: None,
                 sharder: None,
                 pushed_filters: None,
+                aggregation: None,
             },
         )
         .build();
@@ -1213,6 +1214,7 @@ mod tests {
                 filters: None,
                 sharder: None,
                 pushed_filters: None,
+                aggregation: None,
             },
         )
         .build();