--- a/src/daft-logical-plan/src/optimization/rules/shard_scans.rs
+++ b/src/daft-logical-plan/src/optimization/rules/shard_scans.rs
@@ -74,6 +74,8 @@ mod tests {
             )])),
             num_scan_tasks,
             num_rows_per_task: Some(1000),
+            supports_count_pushdown_flag: false,
+            can_absorb_aggregation_flag: false,
         });
 
         let scan_tasks = scan_operator.to_scan_tasks(Pushdowns::default())?;