--- a/src/daft-logical-plan/src/optimization/rules/split_udfs.rs
+++ b/src/daft-logical-plan/src/optimization/rules/split_udfs.rs
@@ -1339,6 +1339,7 @@ mod tests {
                 filters: None,
                 sharder: None,
                 pushed_filters: None,
+                aggregation: None,
             },
         )
         .build();