--- a/src/daft-physical-plan/src/test/mod.rs
+++ b/src/daft-physical-plan/src/test/mod.rs
@@ -11,6 +11,8 @@ pub fn dummy_scan_operator(fields: Vec<Field>) -> ScanOperatorRef {
         schema,
         num_scan_tasks: 1,
         num_rows_per_task: None,
+        supports_count_pushdown_flag: false,
+        can_absorb_aggregation_flag: false,
     }))
 }