--- a/src/daft-scan/src/anonymous.rs
+++ b/src/daft-scan/src/anonymous.rs
@@ -65,6 +65,14 @@ impl ScanOperator for AnonymousScanOperator {
         false
     }
 
+    fn can_absorb_aggregation(&self) -> bool {
+        false
+    }
+
+    fn supports_count_pushdown(&self) -> bool {
+        false
+    }
+
     fn multiline_display(&self) -> Vec<String> {
         let mut lines = vec![
             "AnonymousScanOperator".to_string(),