--- a/src/daft-scan/src/glob.rs
+++ b/src/daft-scan/src/glob.rs
@@ -352,10 +352,6 @@ impl ScanOperator for GlobScanOperator {
         false
     }
 
-    fn can_absorb_aggregation(&self) -> bool {
-        false
-    }
-
     fn supports_count_pushdown(&self) -> bool {
         false
     }