… Cargo.lock
… src/common/scan-info/Cargo.toml
… src/daft-logical-plan/src/optimization/rules/mod.rs
… src/daft-logical-plan/src/test/mod.rs
… src/daft-scan/src/python.rs
… tests/io/lancedb/test_lancedb_reads.py