## Changes Made

<!-- Describe what changes were made and why. Include implementation details if necessary. -->
### 1. Solution Design
#### 1.1 Problem Background
In Daft, performing a `count() `operation on a LanceDB data source requires scanning the entire dataset and pulling the data from the storage layer to the computation layer for aggregation, even if it's just to calculate the total number of rows. This approach is inefficient, especially for very large datasets, as it consumes significant network bandwidth and computational resources.
#### 1.2 Solution
To address this issue, we propose an optimization that pushes down the `count()` aggregation operation to the LanceDB data source. With this solution, Daft can directly leverage LanceDB's metadata or its efficient counting capabilities to obtain the total row count, thus avoiding unnecessary data transfer and computation.
The core idea of this solution is to introduce a new logical optimization rule, `PushDownAggregation`, which identifies `count() `aggregation operations that can be pushed down and converts them into specific scan tasks for the LanceDB data source.
### 2. Code Implementation Logic
#### 2.1 `PushDownAggregation` Optimization Rule
We implemented the `PushDownAggregation` optimization rule in the `src/daft-logical-plan/src/optimization/rules/push_down_aggregation.rs` file. This rule traverses the logical plan tree and matches `Aggregate` nodes.
The rule pushes down the `count()` aggregation operation to the data source when all of the following conditions are met:
**1. No `groupby`**: The aggregation operation does not have a `groupby` clause.
**2. Single Aggregation**: There is only one aggregation expression.
**3. `count` Aggregation**: The aggregation expression is `count()`.
**4. Physical Data Source**: The input to the aggregation operation is a physical data source, not in-memory data or a placeholder.
**5. Data Source Support**: The data source's scanner supports aggregation pushdown, specifically `count` pushdown, and the pushdown `count` mode is supported (currently only `CountMode::All` is supported).
**6. No Filters**: There are no `filters` that have already been pushed down to the data source.
When these conditions are met, the rule creates a new `Source` node and includes the `count()` aggregation information as part of the `pushdowns`. The schema of the new `Source` node is updated to the post-aggregation schema.
#### 2.2 Modifications to `LanceDBScanOperator`
The `LanceDBScanOperator` in the `daft/io/lance/lance_scan.py` file was modified as follows:
**1. Capability Declaration:**
  - The `can_absorb_aggregation()` method returns `True`, indicating that the scanner can absorb aggregation operations.
  - The `supports_count_pushdown()` method returns `True`, indicating that the scanner supports count pushdown.
  - The `supported_count_modes()` method returns `[CountMode.All]`, indicating that the scanner only supports `count(*)` style pushdown.
**2. Task Creation:**
  - The `to_scan_tasks()` method was modified to handle `count` pushdown. When a `count` aggregation pushdown is detected, it creates a `ScanTask` that calls `_lancedb_count_result_function`.
  - The `_lancedb_count_result_function function` is responsible for calling LanceDB's `count_rows()` API to perform the count. In scenarios without filters, this function can directly use the metadata of the LanceDB dataset to efficiently return the row count, thus avoiding costly full data scans.
### 3. Core Benefits
- Improved Query Performance: Pushing down the `count()` computation to LanceDB eliminates the overhead of data transfer and upper-layer aggregation, significantly improving the execution speed of counting queries.
- Optimized Resource Utilization: Significantly reduces the network bandwidth pressure caused by data pulling and saves CPU and memory resources of the computing cluster.
### 4. Limitations and Future Work
#### 4.1 Current Limitations
- count Mode Support: Currently, only `count(*) (CountMode::All)` is supported. `CountMode::Valid (non-null count)` and `CountMode::Null (null count)` are not yet supported.
- `Filter` Limitation: This optimization does not take effect when the query includes a `filter`.
- `groupby`: Aggregation operations with `groupby` are not pushed down.
In the above scenarios, Daft will automatically fall back to the original execution logic, completing the aggregation at the computation layer to ensure the accuracy of the results.
#### 4.2 Future Optimization Directions (TODO)
- Support Pushdown with Filters: Implement the pushdown of `count` operations with filter conditions to the data source.
- Extend `count` Modes: Support modes like `count(column)` by converting them into pushdowns with an `IS NOT NULL` filter.
- Extend to Other Data Sources: Generalize this optimization to other data sources that support efficient counting, such as Iceberg and Hudi.
#### 5. Performance Test Comparison
Data volume: 5000 pieces, reference for test cases: https://github.com/Eventual-Inc/Daft/pull/4874

**Before Optimization**
```
------------------------------------------------------------------------------------------------------------- benchmark: 4 tests ------------------------------------------------------------------------------------------------------------
Name (time in us)                                          Min                    Max                   Mean                StdDev                 Median                   IQR            Outliers         OPS            Rounds  Iterations
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_lance_native_read[count-COUNT(1)]                248.4840 (1.0)       1,158.2470 (1.0)         294.7342 (1.0)         60.6290 (1.0)         287.6330 (1.0)         33.7613 (1.0)         13;13  3,392.8873 (1.0)         369           1
test_duckdb_read_lance[count]                       2,875.4600 (11.57)     3,465.7710 (2.99)      3,078.6922 (10.45)      238.5957 (3.94)      2,959.2820 (10.29)      298.4882 (8.84)          1;0    324.8132 (0.10)          5           1
test_daft_dataframe_read_lance[count-COUNT(1)]     70,618.5060 (284.20)   75,446.5320 (65.14)    72,187.9156 (244.93)   1,905.4424 (31.43)    71,878.5300 (249.90)   1,872.3900 (55.46)         1;0     13.8527 (0.00)          5           1
test_daft_sql_read_lance[count-COUNT(1)]           70,713.5820 (284.58)   72,332.8790 (62.45)    71,581.7112 (242.87)     731.3439 (12.06)    71,935.9190 (250.10)   1,271.6818 (37.67)         2;0     13.9700 (0.00)          5           1
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```

**After Optimization**
```
--------------------------------------------------------------------------------------------------------- benchmark: 4 tests --------------------------------------------------------------------------------------------------------
Name (time in us)                                         Min                   Max                  Mean              StdDev                Median                 IQR            Outliers         OPS            Rounds  Iterations
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_lance_native_read[count-COUNT(1)]               243.4360 (1.0)      2,890.7980 (1.0)        288.3970 (1.0)       99.7140 (1.0)        282.7590 (1.0)       30.4350 (1.0)         11;25  3,467.4419 (1.0)        1572           1
test_duckdb_read_lance[count]                      3,053.7880 (12.54)    3,503.1090 (1.21)     3,275.2938 (11.36)    175.0964 (1.76)     3,300.6170 (11.67)    264.3005 (8.68)          2;0    305.3161 (0.09)          5           1
test_daft_sql_read_lance[count-COUNT(1)]           6,075.9390 (24.96)    6,461.6360 (2.24)     6,240.9604 (21.64)    146.2203 (1.47)     6,259.2170 (22.14)    182.1147 (5.98)          2;0    160.2317 (0.05)          5           1
test_daft_dataframe_read_lance[count-COUNT(1)]     6,362.0170 (26.13)    8,084.4090 (2.80)     6,890.7954 (23.89)    705.4063 (7.07)     6,671.3280 (23.59)    825.6568 (27.13)         1;0    145.1211 (0.04)          5           1
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to `docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130 for docs review)