@huleilei 2025-08-14
> This looks pretty sound to me. Thanks @huleilei
> 
> My biggest question is if we can later extend this functionality to parquet as well. Right now parquet count() does a full scan as seen here #2298. This PR seems to lay a lot of the ground work for whats needed to support this in parquet as well.

Yes. Other data sources also need to consider optimizing count_row. My next PR will apply this feature to the Parquet file. Thanks